rebase, remove chlo and clang-format
