+++ lib/Conversion/TorchToMhlo/CMakeLists.txt
@@ -8,7 +8,6 @@ add_mlir_conversion_library(TorchMLIRTorchToMhlo
 
   DEPENDS
   MhloDialect
-  ChloDialect
   TorchMLIRConversionPassIncGen
 
   LINK_COMPONENTS
@@ -18,7 +17,6 @@ add_mlir_conversion_library(TorchMLIRTorchToMhlo
   MLIRIR
   MLIRPass
   MhloDialect
-  ChloDialect
   TorchMLIRTorchDialect
 )
